fix: mark attendance error + home page + mark attendance for all class
