+++ src/components/attendance/AttendanceComponent.js
@@ -3,13 +3,15 @@ import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import HdrAutoIcon from "@mui/icons-material/HdrAuto";
 import CircleIcon from "@mui/icons-material/Circle";
 import {
-  IconButton,
   VStack,
   Text,
   HStack,
   Box,
   Pressable,
   Actionsheet,
+  Stack,
+  FlatList,
+  Button,
 } from "native-base";
 import * as attendanceServiceRegistry from "../../services/attendanceServiceRegistry";
 import manifest from "../../modules/attendance/manifest.json";
@@ -20,6 +22,9 @@ import WatchLaterIcon from "@mui/icons-material/WatchLater";
 import moment from "moment";
 import Card from "../students/Card";
 import { CircleOutlined } from "@mui/icons-material";
+import Header from "../Header";
+import { Link } from "react-router-dom";
+import IconByName from "../IconByName";
 
 export function weekDaysPageWise(weekPage, today) {
   let date = new Date();
@@ -154,6 +159,413 @@ export const GetIcon = ({ status, _box, color }) => {
   return icon;
 };
 
+export const MultipalAttendance = ({
+  students,
+  attendance,
+  getAttendance,
+  setLoding,
+  setAllAttendanceStatus,
+  allAttendanceStatus,
+  classObject,
+  isEditDisabled,
+  setIsEditDisabled,
+}) => {
+  const { t } = useTranslation();
+  const [showModal, setShowModal] = useState(false);
+  const teacherId = sessionStorage.getItem("id");
+  const status = manifest?.status ? manifest?.status : [];
+
+  const getStudentsAttendance = (e) => {
+    return students
+      .map((item) => {
+        return attendance
+          .slice()
+          .reverse()
+          .find(
+            (e) =>
+              e.date === moment().format("Y-MM-DD") && e.studentId === item.id
+          );
+      })
+      .filter((e) => e);
+  };
+
+  const countReport = ({ gender, attendanceType, type }) => {
+    let attendanceAll = getStudentsAttendance();
+
+    let studentIds = students.map((e) => e.id);
+    if (gender && [t("BOYS"), t("GIRLS")].includes(gender)) {
+      studentIds = students
+        .filter(
+          (e) =>
+            e.gender ===
+            (gender === t("BOYS")
+              ? "Male"
+              : gender === t("GIRLS")
+              ? "Female"
+              : "")
+        )
+        .map((e) => e.id);
+    }
+    if (type === "Total" && gender === t("TOTAL")) {
+      return studentIds.length;
+    } else if (type === "Total" && [t("BOYS"), t("GIRLS")].includes(gender)) {
+      let studentIds1 = studentIds.filter(
+        (e) =>
+          !attendanceAll
+            .filter((e) => studentIds.includes(e?.studentId))
+            .map((e) => e.studentId)
+            .includes(e)
+      );
+
+      return (
+        attendanceAll.filter((e) => studentIds.includes(e?.studentId)).length +
+        studentIds1.length
+      );
+    } else if (attendanceType === "Unmarked" && gender === t("TOTAL")) {
+      let studentIds1 = attendanceAll.filter(
+        (e) =>
+          studentIds.includes(e.studentId) && e.attendance !== attendanceType
+      );
+      return Math.abs(studentIds.length - studentIds1.length);
+    } else if (type === "Unmarked" || attendanceType === "Unmarked") {
+      let studentIds1 = attendanceAll.filter((e) =>
+        studentIds.includes(e.studentId)
+      );
+
+      if (attendanceType === "Unmarked") {
+        studentIds1 = attendanceAll.filter(
+          (e) =>
+            studentIds.includes(e?.studentId) && e.attendance !== attendanceType
+        );
+      }
+      return Math.abs(studentIds.length - studentIds1.length);
+    } else {
+      return attendanceAll.filter(
+        (e) =>
+          studentIds.includes(e?.studentId) && e.attendance === attendanceType
+      ).length;
+    }
+  };
+
+  const markAllAttendance = async () => {
+    setLoding(true);
+    if (typeof students === "object") {
+      let ctr = 0;
+      let attendanceAll = getStudentsAttendance();
+
+      students.forEach((item, index) => {
+        let attendanceObject = attendanceAll.find(
+          (e) => item.id === e.studentId
+        );
+        let result = null;
+        if (attendanceObject?.id) {
+          if (attendanceObject.attendance !== "Present") {
+            result = attendanceServiceRegistry
+              .update(
+                {
+                  id: attendanceObject.id,
+                  attendance: "Present",
+                },
+                {
+                  headers: {
+                    Authorization: "Bearer " + sessionStorage.getItem("token"),
+                  },
+                }
+              )
+              .then((e) => {
+                if (getAttendance) {
+                  getAttendance();
+                }
+              });
+          } else {
+            result = "alreadyPresent";
+          }
+        } else {
+          result = attendanceServiceRegistry.create(
+            {
+              studentId: item.id,
+              date: moment().format("Y-MM-DD"),
+              attendance: "Present",
+              attendanceNote: "Test",
+              classId: item.currentClassID,
+              subjectId: "History",
+              teacherId: teacherId,
+            },
+            {
+              headers: {
+                Authorization: "Bearer " + sessionStorage.getItem("token"),
+              },
+            }
+          );
+        }
+
+        setTimeout(async (e) => {
+          if (result && result === "alreadyPresent") {
+            setAllAttendanceStatus({
+              ...allAttendanceStatus,
+              success: parseInt(index + 1) + " Already Present",
+            });
+          } else if (result) {
+            setAllAttendanceStatus({
+              ...allAttendanceStatus,
+              success: parseInt(index + 1) + " success",
+            });
+          } else {
+            setAllAttendanceStatus({
+              ...allAttendanceStatus,
+              fail: parseInt(index + 1) + " fail",
+            });
+          }
+          ctr++;
+          if (ctr === students.length) {
+            setAllAttendanceStatus({});
+            setLoding(false);
+            await getAttendance();
+          }
+        }, index * 900);
+      });
+    }
+  };
+
+  const PopupActionSheet = () => {
+    return (
+      <>
+        <Actionsheet isOpen={showModal} onClose={() => setShowModal(false)}>
+          <Actionsheet.Content bg="coolGray.500">
+            <Header
+              isDisabledAppBar={true}
+              _box={{ bg: "coolGray.500", py: 0 }}
+              icon="AssignmentTurnedIn"
+              heading={t("ATTENDANCE")}
+              _heading={{ fontSize: "xl" }}
+              subHeadingComponent={
+                <Link
+                  to={"/students/class/" + classObject.id}
+                  style={{ color: "rgb(63, 63, 70)", textDecoration: "none" }}
+                >
+                  <Box
+                    rounded="full"
+                    borderColor="coolGray.200"
+                    borderWidth="1"
+                    bg="white"
+                    px={1}
+                  >
+                    <HStack
+                      space="4"
+                      justifyContent="space-between"
+                      alignItems="center"
+                    >
+                      <IconByName size="sm" name="Group" />
+                      <Text fontSize={"lg"}>{classObject?.title ?? ""}</Text>
+                      <IconByName size="sm" name="ArrowForwardIos" />
+                    </HStack>
+                  </Box>
+                </Link>
+              }
+            />
+          </Actionsheet.Content>
+          <Box bg="coolGray.100" width={"100%"}>
+            <Box bg="white" p="2">
+              <HStack justifyContent="space-between" alignItems="center">
+                <Text>{t("ATTENDANCE_SUMMARY")}</Text>
+                <Text>
+                  {t("TODAY")}: {moment().format("ddd DD, MMM")}
+                </Text>
+              </HStack>
+            </Box>
+            <Box p={4}>
+              <Stack space={3}>
+                <Box borderWidth={1} p="2" borderColor="gray.500" bg="gray.50">
+                  <FlatList
+                    data={["heade", t("BOYS"), t("GIRLS"), t("TOTAL")]}
+                    renderItem={({ item, index }) => (
+                      <HStack
+                        alignItems={"center"}
+                        space={2}
+                        justifyContent={"space-between"}
+                      >
+                        {item === "heade" ? (
+                          <Text></Text>
+                        ) : (
+                          <Text fontSize="2xl">{item}</Text>
+                        )}
+                        {item === "heade" ? (
+                          <HStack alignItems={"center"} space={2}>
+                            {status.map((item, index) => {
+                              return (
+                                <GetIcon
+                                  key={index}
+                                  status={item}
+                                  _box={{ p: 2, minW: "55", maxW: "55" }}
+                                />
+                              );
+                            })}
+                            {!status.includes("Unmarked") ? (
+                              <GetIcon
+                                status={"Today"}
+                                _box={{ p: 2, minW: "55", maxW: "55" }}
+                              />
+                            ) : (
+                              <></>
+                            )}
+                            <Text fontSize="2xl" minW={"55"} maxW={"55"}>
+                              {t("TOTAL")}
+                            </Text>
+                          </HStack>
+                        ) : (
+                          <HStack
+                            alignItems={"center"}
+                            space={2}
+                            width={"252px"}
+                          >
+                            {status.map((subItem, index) => {
+                              return (
+                                <Text
+                                  key={index}
+                                  fontSize="2xl"
+                                  minW={"55"}
+                                  maxW={"55"}
+                                  textAlign={"center"}
+                                >
+                                  {countReport({
+                                    gender: item,
+                                    attendanceType: subItem,
+                                  })}
+                                </Text>
+                              );
+                            })}
+                            {!status.includes("Unmarked") ? (
+                              <Text
+                                fontSize="2xl"
+                                minW={"55"}
+                                maxW={"55"}
+                                textAlign={"center"}
+                              >
+                                {countReport({
+                                  type: "Unmarked",
+                                  gender: item,
+                                })}
+                              </Text>
+                            ) : (
+                              <></>
+                            )}
+                            <Text
+                              fontSize="2xl"
+                              minW={"55"}
+                              maxW={"55"}
+                              textAlign={"center"}
+                            >
+                              {countReport({
+                                type: "Total",
+                                gender: item,
+                              })}
+                            </Text>
+                          </HStack>
+                        )}
+                      </HStack>
+                    )}
+                    keyExtractor={(item, index) => index}
+                  />
+                </Box>
+                <VStack>
+                  <Box
+                    borderWidth={1}
+                    p="2"
+                    borderColor="gray.500"
+                    bg="gray.50"
+                  >
+                    <Text>
+                      <Text>100% {t("THIS_WEEK")}: </Text>
+                      <Text pr={1}>
+                        {students.map((e) => e.fullName).join(", ")}
+                      </Text>
+                    </Text>
+                  </Box>
+                </VStack>
+                <VStack space={4} width={"100%"} alignItems={"center"}>
+                  <Link
+                    to={"/classes/" + classObject.id}
+                    style={{ color: "rgb(63, 63, 70)", textDecoration: "none" }}
+                  >
+                    <Box
+                      rounded={"full"}
+                      background="coolGray.600"
+                      px={6}
+                      py={2}
+                    >
+                      <Text color="coolGray.50">{t("SEE_FULL_REPORT")}</Text>
+                    </Box>
+                  </Link>
+                  <Button
+                    variant="ghost"
+                    rounded={"full"}
+                    colorScheme="gray"
+                    background="coolGray.300"
+                    px={6}
+                    width={"100%"}
+                    onPress={(e) => setShowModal(false)}
+                  >
+                    {t("CLOSE")}
+                  </Button>
+                </VStack>
+              </Stack>
+            </Box>
+          </Box>
+        </Actionsheet>
+      </>
+    );
+  };
+
+  return (
+    <Stack position={"sticky"} bottom={0} width={"100%"}>
+      <Box p="2" bg="coolGray.400">
+        <VStack space={3} alignItems={"center"}>
+          <Text textAlign={"center"}>
+            {t("ATTENDANCE_WILL_AUTOMATICALLY_SUBMIT")}
+          </Text>
+          {!isEditDisabled ? (
+            <HStack alignItems={"center"} space={4}>
+              <Button
+                variant="ghost"
+                rounded={"full"}
+                colorScheme="gray"
+                background="coolGray.200"
+                px={4}
+                onPress={markAllAttendance}
+              >
+                {t("MARK_ALL_PRESENT")}
+              </Button>
+              <Button
+                borderRadius="50"
+                colorScheme="gray"
+                background="coolGray.600"
+                px={6}
+                onPress={(e) => setShowModal(true)}
+              >
+                {t("SAVE")}
+              </Button>
+            </HStack>
+          ) : (
+            <HStack alignItems={"center"} space={4}>
+              <Button
+                variant="ghost"
+                rounded={"full"}
+                colorScheme="gray"
+                background="coolGray.200"
+                px={4}
+                onPress={(e) => setIsEditDisabled(false)}
+              >
+                {t("EDIT")}
+              </Button>
+            </HStack>
+          )}
+        </VStack>
+      </Box>
+      <PopupActionSheet />
+    </Stack>
+  );
+};
+
 const AttendanceComponent = ({
   today,
   weekPage,
@@ -164,7 +576,7 @@ const AttendanceComponent = ({
   withApigetAttendance,
   getAttendance,
   _card,
-  editState,
+  isEditDisabled,
 }) => {
   const { t } = useTranslation();
   const todayDate = new Date();
@@ -176,27 +588,38 @@ const AttendanceComponent = ({
   const [loding, setLoding] = useState({});
   const status = manifest?.status ? manifest?.status : [];
 
-  useEffect(async () => {
+  useEffect(() => {
     setWeekDays(weekDaysPageWise(weekPage, today));
-    if (withApigetAttendance) {
-      setAttendance(
-        await GetAttendance({
-          studentId: {
-            eq: student.id,
-          },
-          classId: {
-            eq: student.currentClassID,
-          },
-          teacherId: {
-            eq: teacherId,
-          },
-        })
-      );
-    } else if (attendanceProp) {
-      setAttendance(attendanceProp);
+    async function getData() {
+      if (withApigetAttendance) {
+        setAttendance(
+          await GetAttendance({
+            studentId: {
+              eq: student.id,
+            },
+            classId: {
+              eq: student.currentClassID,
+            },
+            teacherId: {
+              eq: teacherId,
+            },
+          })
+        );
+      } else if (attendanceProp) {
+        setAttendance(attendanceProp);
+      }
+      setLoding({});
     }
-    setLoding({});
-  }, [weekPage, attendanceProp, withApigetAttendance && !showModal]);
+    getData();
+  }, [
+    weekPage,
+    attendanceProp,
+    withApigetAttendance && !showModal,
+    student.id,
+    student.currentClassID,
+    teacherId,
+    today,
+  ]);
 
   const formatDate = (date) => {
     var d = new Date(date),
@@ -302,28 +725,13 @@ const AttendanceComponent = ({
                 </Actionsheet.Item>
               );
             })}
-            {/* <Actionsheet.Item p={1}>
-              <Pressable onPress={(e) => setShowModal(false)}>
-                <HStack alignItems="center" space={2}>
-                  <IconButton
-                    size="md"
-                    py="2"
-                    color={"gray.400"}
-                    icon={<CircleIcon fontSize="large" />}
-                  />
-                  <Text color="coolGray.800" bold fontSize="lg">
-                    {t("Unmarked")}
-                  </Text>
-                </HStack>
-              </Pressable>
-            </Actionsheet.Item> */}
           </Actionsheet.Content>
         </Actionsheet>
       </>
     );
   };
 
-  const WeekDaysComponent = ({ weekDays, isIconSizeSmall, editState }) => {
+  const WeekDaysComponent = ({ weekDays, isIconSizeSmall, isEditDisabled }) => {
     return weekDays.map((day, subIndex) => {
       let dateValue = formatDate(day);
       let attendanceItem = attendance
@@ -404,7 +812,7 @@ const AttendanceComponent = ({
             )}
             <TouchableHighlight
               onPress={(e) => {
-                if (editState) {
+                if (!isEditDisabled) {
                   markAttendance({
                     attendanceId: attendanceItem?.id ? attendanceItem.id : null,
                     date: dateValue,
@@ -414,7 +822,7 @@ const AttendanceComponent = ({
                 }
               }}
               onLongPress={(event) => {
-                if (editState) {
+                if (!isEditDisabled) {
                   setAttendanceObject({
                     attendanceId: attendanceItem?.id ? attendanceItem.id : null,
                     date: dateValue,
@@ -455,7 +863,7 @@ const AttendanceComponent = ({
               <WeekDaysComponent
                 weekDays={weekDays}
                 isIconSizeSmall={true}
-                editState={editState}
+                isEditDisabled={isEditDisabled}
               />
             ) : (
               false
@@ -465,7 +873,10 @@ const AttendanceComponent = ({
       </Box>
       {!today ? (
         <HStack justifyContent="space-between" alignItems="center">
-          <WeekDaysComponent weekDays={weekDays} editState={editState} />
+          <WeekDaysComponent
+            weekDays={weekDays}
+            isEditDisabled={isEditDisabled}
+          />
         </HStack>
       ) : (
         <></>
