+++ src/shiksha-os/home.js
@@ -1,15 +1,54 @@
-import React from "react";
-import { Text, Box, Stack, VStack, HStack, Button } from "native-base";
+import React, { useEffect, useState } from "react";
+import { Text, Box, Stack, VStack, HStack } from "native-base";
 import manifest from "./manifest";
 import Menu from "../components/Menu";
 import Header from "../components/Header";
 import { useTranslation } from "react-i18next";
+import * as classServiceRegistry from "../shiksha-os/services/classServiceRegistry";
+// import * as studentServiceRegistry from "../shiksha-os/services/studentServiceRegistry";
+// import * as attendanceServiceRegistry from "../services/attendanceServiceRegistry";
 
 // Start editing here, save and see your changes.
 export default function Home() {
   const menus = manifest.menus.main;
   const { t } = useTranslation();
   const firstName = sessionStorage.getItem("firstName");
+  const [classes, setClasses] = useState([]);
+  const [students, setStudents] = useState([]);
+  const authId = sessionStorage.getItem("id");
+  const [attendance, setAttendance] = useState([]);
+
+  useEffect(() => {
+    let ignore = false;
+    const getData = async () => {
+      if (!ignore) {
+        let resultClasses = await classServiceRegistry.getAll({
+          filters: {
+            teacherId: {
+              eq: authId,
+            },
+          },
+        });
+        setClasses(resultClasses);
+        // let redultStudents = await studentServiceRegistry.getAll({
+        //   filters: {
+        //     currentClassID: {
+        //       in: resultClasses.map((e) => e.id),
+        //     },
+        //   },
+        // });
+        // setStudents(redultStudents);
+        // const attendanceData = await attendanceServiceRegistry.getAll({
+        //   classId: {
+        //     in: redultStudents.map((e) => e.id),
+        //   },
+        // });
+        // console.log({ attendanceData });
+        // setAttendance(attendanceData);
+      }
+    };
+    getData();
+  }, [authId]);
 
   return (
     <>
@@ -32,20 +71,15 @@ export default function Home() {
               <Text color="green.700" bold={true}>
                 {t("TODAY")}
               </Text>
-              <Button
-                rounded={50}
-                colorScheme="default"
-                bg={"coolGray.700"}
-                color={"coolGray.100"}
-                size="container"
-                px={3}
-              >
-                {t("TIME_TABLE")}
-              </Button>
             </HStack>
             <Stack>
               <VStack>
-                <Text>{t("YOU_HAVE_3_CLASSES")}</Text>
+                <Text>
+                  {t("YOU_HAVE_3_CLASSES").replace(
+                    "3",
+                    classes?.length ? classes?.length : "3"
+                  )}
+                </Text>
                 <Text>{t("STUDENT_PROFILE_IN_COMPLETE")}</Text>
                 <Text>{t("NEW_ACTIVITY_ADDED_TO_SCHOOL")}</Text>
               </VStack>
@@ -56,16 +90,6 @@ export default function Home() {
               <Text color="green.700" bold={true}>
                 {t("THIS_WEEK")}
               </Text>
-              <Button
-                rounded={50}
-                colorScheme="default"
-                bg={"coolGray.700"}
-                color={"coolGray.100"}
-                size="container"
-                px={3}
-              >
-                {t("CALENDAR")}
-              </Button>
             </HStack>
             <HStack space={2}>
               <Text>{t("2_SCHOOL_ACTIVITIES_1_OFFICIAL_VISIT")}</Text>
