+++ src/shiksha-os/lang/samagra-shiksha.en-gb.json
@@ -83,5 +83,6 @@
   "CLASS_CHOOE_ATTENDANCE": "Please choose class / group to mark attendance",
   "SELECT_CLASS": "Select class / group",
   "SELECT_SUBJECT": "Select subject",
-  "EDIT": "Edit"
+  "EDIT": "Edit",
+  "CLOSE": "Close"
 }
