+++ src/shiksha-os/lang/samagra-shiksha.hi-in.json
@@ -83,5 +83,6 @@
   "CLASS_CHOOE_ATTENDANCE": "उपस्थिति दर्ज करने के लिए कृपया कक्षा/समूह चुनें",
   "SELECT_CLASS": "वर्ग / समूह का चयन करें",
   "SELECT_SUBJECT": "विषय चुनें",
-  "EDIT": "संपादित करें"
+  "EDIT": "संपादित करें",
+  "CLOSE": "बंद करे"
 }
